move to helper function and use meshsdk lookup
